feat: bundle prebuilt PJSIP (pjsua2) and port to real bindings
Build PJSIP 2.16 + pjsua2 (G.711/G.722/Opus, OpenSSL/TLS) via
pjsip-android-builder and commit the artifacts so the app builds with SIP
enabled out of the box: pjsua2.jar plus libpjsua2.so + libc++_shared.so for
arm64-v8a, armeabi-v7a and x86_64 (native .so stored in Git LFS).

Port the app and the compile-time stub from the hand-written enum API
(MediaFrameType, .swigValue()/.swigToEnum()) to real pjsua2's int-constant
enums, so both the real and stub build paths compile. Unit tests pass.

+++ app/src/main/java/com/magick/gsm2sip/audio/PjBytes.kt
@@ -6,7 +6,7 @@ import org.pjsip.pjsua2.ByteVector
 fun ByteVector.toByteArray(): ByteArray {
     val n = size.toInt()
     val out = ByteArray(n)
-    for (i in 0 until n) out[i] = get(i)
+    for (i in 0 until n) out[i] = get(i).toByte()
     return out
 }
 
@@ -17,7 +17,7 @@ fun ByteVector.toByteArray(): ByteArray {
 fun ByteArray.toByteVectorSafe(targetBytes: Int): ByteVector {
     val vec = ByteVector()
     val n = minOf(size, targetBytes)
-    for (i in 0 until n) vec.add(this[i])
-    for (i in n until targetBytes) vec.add(0.toByte())
+    for (i in 0 until n) vec.add(this[i].toShort())
+    for (i in n until targetBytes) vec.add(0.toShort())
     return vec
 }
